requirements.txt
# LayerX Steganographic Security Framework - Member A Requirements
# Install with: pip install -r requirements.txt

# Core cryptographic operations
pycryptodome>=3.18.0

# Image processing and mathematical operations  
numpy>=1.24.0
opencv-python>=4.8.0
pywavelets>=1.4.1
scikit-image>=0.21.0

# Signal processing for DCT/IDCT
scipy>=1.10.0

# Development and testing
pytest>=7.4.0